refactor: Reorganize core files into proper layers (#203)
- Moved `AccessTokenInterceptor` and `TokenAuthenticator` to `data.auth`.
- Relocated custom exceptions from `core.error` -> `core.exceptions`.
- Moved `MultipartUtils.kt` to `common.util`.
- Updated all import paths across the project to match the new structure.
