+++ app/src/main/java/com/example/rentit/core/exceptions/ConflictException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.core.error
+package com.example.rentit.core.exceptions
 
 class ConflictException(message: String? = null): Exception(message)
