+++ app/src/main/java/com/example/rentit/core/exceptions/NotFoundException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.core.error
+package com.example.rentit.core.exceptions
 
 class NotFoundException(message: String? = null): Exception(message)
