+++ app/src/main/java/com/example/rentit/core/exceptions/ServerErrorException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.core.error
+package com.example.rentit.core.exceptions
 
 class ServerErrorException(message: String? = null): Exception(message)
