+++ app/src/main/java/com/example/rentit/core/exceptions/UnauthorizedException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.core.error
+package com.example.rentit.core.exceptions
 
 class UnauthorizedException(message: String? = null): Exception(message)
