+++ app/src/main/java/com/example/rentit/core/exceptions/UnknownException.kt
@@ -1,3 +1,3 @@
-package com.example.rentit.core.error
+package com.example.rentit.core.exceptions
 
 class UnknownException(message: String? = null): Exception(message)
