+++ app/src/main/java/com/example/rentit/presentation/main/MainViewModel.kt
@@ -2,7 +2,7 @@ package com.example.rentit.presentation.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.rentit.core.error.UnauthorizedException
+import com.example.rentit.core.exceptions.UnauthorizedException
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
